% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{cluster_determine}
\alias{cluster_determine}
\title{Determine number of clusters
http://www.bagualu.net/wordpress/wp-content/uploads/2015/10/A_Handbook_of_Statistical_Analyses_Using_R__Second_Edition.pdf
page 329}
\usage{
cluster_determine(input, init_claster_num = 15)
}
\arguments{
\item{input}{- input object}

\item{init_claster_num}{- initial number of clusters}
}
\value{
count_claster - evaluated optimal count of clusters
}
\description{
Determine number of clusters
http://www.bagualu.net/wordpress/wp-content/uploads/2015/10/A_Handbook_of_Statistical_Analyses_Using_R__Second_Edition.pdf
page 329
}
