% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{create_nan_vars}
\alias{create_nan_vars}
\title{Creates boolean vars for NaN observations - 1 if Nan, 0 - otherwise}
\usage{
create_nan_vars(.dt, range = 1:ncol(.dt), threshold = 0.05, verbose = T)
}
\arguments{
\item{.dt}{- input data.table to which the transformation to be applied}

\item{range}{- range of columns of dt to which the transformation to be applied}

\item{threshold}{- a NaN share of input population, sufficient for creation of a boolean var}

\item{verbose}{- indicates if detailed messages being shown in console}
}
\value{
dt - output data.table with boolean vars for NaNs created
}
\description{
Creates boolean vars for NaN observations - 1 if Nan, 0 - otherwise
}
