% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{dummy_transform}
\alias{dummy_transform}
\title{Transforms text variables of input data.table within the given range to a set of dummy variables}
\usage{
dummy_transform(.dt, range = 1:ncol(.dt), delete_origin = T)
}
\arguments{
\item{.dt}{- input data.table to which the transformation to be applied}

\item{range}{- range of columns of dt to which the transformation to be applied}

\item{delete_origin}{- indicates if origin text varibles to be deleted}
}
\value{
dt - output data.table with dummy variables for text
}
\description{
Transforms text variables of input data.table within the given range to a set of dummy variables
}
