% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{mode_value}
\alias{mode_value}
\title{Get the ‘mode’ (a kind of value) of an R object.}
\usage{
mode_value(x, na_rm = T)
}
\arguments{
\item{x}{- input object}

\item{na_rm}{- indicates if NA, NaN, Inf values should be removed. Default is TRUE}
}
\value{
res - output value of a statistic mode
}
\description{
Get the ‘mode’ (a kind of value) of an R object.
}
