% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Removes outliers from the chosen characteristics of input data table
using a formula Lower quantile - H, Upper quantile + H, where H = coeff * (interquantile distance)}
\usage{
remove_outliers(
  .dt,
  range = 1:ncol(.dt),
  probs = c(0.25, 0.75),
  coeff = 1.5,
  replace = "mean",
  create_variable = 0.05,
  verbose = T
)
}
\arguments{
\item{.dt}{- input data.table to which the transformation to be applied}

\item{range}{- range of columns of dt to which the transformation to be applied}

\item{probs}{- a vector of 2 values of lower an upper quantiles}

\item{coeff}{- coefficient for interquartile distance multiplication}

\item{replace}{- indicates an approach for outlier replacement, can be one of 'mean', 'na', 'maxmin'}

\item{create_variable}{- indicates a threshold for a ratio of outlier to full population, starting from which an additional variable will be created to store outlier values}

\item{verbose}{- indicates if detailed messages being shown in console}
}
\value{
dt - output data.table with boolean vars for NaNs created
}
\description{
Removes outliers from the chosen characteristics of input data table
using a formula Lower quantile - H, Upper quantile + H, where H = coeff * (interquantile distance)
}
