% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{replace_nan}
\alias{replace_nan}
\title{Makes data replacement (for NA, NAN, NULL, INF elements).
For numeric parameters for replacement is used a median, for characters - mode}
\usage{
replace_nan(overall_sample_num)
}
\arguments{
\item{overall_sample_num}{- input object to which the transformation to be applied}
}
\value{
overall_sample_mutated - output object with dummy variables for text
}
\description{
Makes data replacement (for NA, NAN, NULL, INF elements).
For numeric parameters for replacement is used a median, for characters - mode
}
