% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{scale_custom}
\alias{scale_custom}
\title{Scale input dataset.}
\usage{
scale_custom(input, range = 1:ncol(input))
}
\arguments{
\item{input}{- input object}

\item{range}{- range of columns to which the transformation to be applied}
}
\value{
res - output object
}
\description{
Scale input dataset.
}
